Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    int i;
-    scanf("%d",&i);
+   float i;
+    scanf("%f",&i);
     float c;
     c = 3.14*(i*i);
     printf("Area: %.2f",c);
